requirements.txt
ccxt>=4.1.13
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-telegram-bot==20.6
pandas>=2.1.3
numpy>=1.26.2
ta>=0.11.0
TA-Lib>=0.4.28
orjson>=3.9.10
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
pydantic==2.5.2
pytest==7.4.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
numba>=0.58.1  # Pour l'accélération des calculs numpy
pandas-ta>=0.3.14b  # Indicateurs techniques additionnels
scipy>=1.11.4  # Pour les calculs scientifiques avancés
matplotlib>=3.7.1
pytz>=2023.3
pytrends>=4.9.0  # API non officielle pour Google Trends
requests>=2.28.0
# Machine Learning et IA
scikit-learn>=1.4.0  # Modèles ML classiques
river>=0.21.0  # Apprentissage incrémental
torch>=2.2.0  # PyTorch pour le deep learning
torchvision>=0.17.0  # Outils de vision par ordinateur PyTorch
torchaudio>=2.2.0  # Traitement audio PyTorch
lightgbm>=4.3.0  # Gradient Boosting optimisé
joblib>=1.3.2  # Parallélisation des calculs ML
optuna>=3.5.0  # Optimisation automatique des hyperparamètres

# Dashboard et Visualisation
dash>=2.14.2  # Framework pour dashboard interactif
plotly>=5.18.0  # Visualisations interactives
dash-bootstrap-components>=1.5.0  # Composants Bootstrap pour Dash
dash-daq>=0.5.0  # Composants analytiques avancés

# Communication Réseau et APIs
httpx>=0.26.0  # Client HTTP/2 moderne et asynchrone
websockets>=12.0  # WebSocket client/server
aiohttp>=3.9.1  # Framework asynchrone
yarl>=1.9.4  # URL parsing optimisé
cryptofeed>=2.4.0  # Agrégateur de flux crypto temps réel
cryptopanic-client>=0.3.0  # Client CryptoPanic API

# Enrichissement et Analyse Sémantique
beautifulsoup4>=4.12.2  # Web scraping
lxml>=5.1.0  # Parser XML/HTML rapide
openai>=1.12.0  # API OpenAI pour GPT
newspaper3k>=0.2.8  # Extraction d'articles
textblob>=0.17.1  # Analyse de sentiment
feedparser>=6.0.11  # Parsing RSS/Atom feeds

# Optimisation et Performance
ujson>=5.9.0  # JSON ultra-rapide
cachetools>=5.3.2  # Mise en cache optimisée
prometheus-client>=0.19.0  # Métriques et monitoring
aiodns>=3.1.1  # DNS asynchrone
brotli>=1.1.0  # Compression haute performance
